Don't return a 400 for `/`
Currently this exporter doesn't behave like other Prometheus exporters, which return a 200 for a `GET HTTP/1.1 /` request. It seems that this was intentional so as to not leak data unintentionally, but I don't personally think that this is a valid reason to deviate from the convention. Specifically, if you don't intend for the Prometheus Varnish exporter to be accessed externally then you should restrict access using security groups, a firewall or `iptables`.

In our use case, we are using `/` as a health check for the exporter itself. This health check is currently failing because the Varnish exporter returns a `400 Bad Request` rather than `200 OK`.